[feature](folder-structure)update folder structure property and property details page
