+++ src/Features/Properties/Components/PropertiesCard/PropertiesCard.jsx
@@ -11,11 +11,14 @@ import 'swiper/css';
 import 'swiper/css/pagination';
 import { Pagination, Autoplay } from 'swiper/modules';
 // for Add to Favorite
-import useAuth from '../../Hooks/useAuth';
-import useAxios from '../../Hooks/useAxios';
+// import useAuth from '../../Hooks/useAuth';
+import useAuth from '../../../../Hooks/useAuth';
+// import useAxios from '../../Hooks/useAxios';
+import useAxios from '../../../../Hooks/useAxios';
 import toast from 'react-hot-toast';
 import Swal from 'sweetalert2';
-import useFavorite from '../../Hooks/useFavorite';
+// import useFavorite from '../../Hooks/useFavorite';
+import useFavorite from '../../../../Hooks/useFavorite';
 
 const PropertiesCard = ({ card }) => {
   const {
@@ -47,7 +50,7 @@ const PropertiesCard = ({ card }) => {
       };
       instance
         .post('/property-favorite/add-favorite', favoriteItem)
-        .then((res) => {
+        .then(res => {
           if (res?.data?.status === 'success') {
             toast.success(`${propertyTitle} added to your favorite`);
             refetch();
@@ -62,7 +65,7 @@ const PropertiesCard = ({ card }) => {
         confirmButtonColor: '#3085d6',
         cancelButtonColor: '#d33',
         confirmButtonText: 'Yes, login!',
-      }).then((result) => {
+      }).then(result => {
         if (result.isConfirmed) {
           navigate('/login', { state: { from: location } });
         }
