+++ src/Features/Properties/Components/PropertiesCard/PropertiesCardList.jsx
