+++ src/Features/Properties/Properties.css
