+++ src/Features/Properties/PropertiesStyle.css
