+++ src/Features/PropertiesDetails/Components/Buyer Inquiry Form/Form.css
@@ -1,0 +1,33 @@
+.slider {
+  -webkit-appearance: none;
+  width: 100%;
+  height: 25px;
+  border-radius: 9999px;
+  background: #d3d3d3;
+  outline: none;
+  opacity: 0.7;
+  -webkit-transition: 0.2s;
+  transition: opacity 0.2s;
+}
+
+.slider:hover {
+  opacity: 1;
+}
+
+.slider::-webkit-slider-thumb {
+  -webkit-appearance: none;
+  appearance: none;
+  border-radius: 9999px;
+  width: 25px;
+  height: 25px;
+  background: #eb6753;
+  cursor: pointer;
+}
+
+.slider::-moz-range-thumb {
+  width: 25px;
+  height: 25px;
+  border-radius: 9999px;
+  background: #eb6753;
+  cursor: pointer;
+}
