+++ src/Features/PropertiesDetails/Components/QRCode/QRcode.jsx
