+++ src/Features/PropertiesDetails/Components/RecommendedProperty/RecommendedCard/RecommendedCard.jsx
