+++ src/Layout/Main.jsx
@@ -1,7 +1,7 @@
 import { Outlet } from 'react-router-dom';
 import Navbar from '../Components/Shared/Navbar/Navbar';
 import Footer from '../Components/Shared/Footer/Footer';
-import TopButton from '../Components/Shared/TopButton/TopButton';
+// import TopButton from '../Components/Shared/TopButton/TopButton';
 import { ThemeProvider } from '../Providers/ThemeContext';
 import { useEffect, useState } from 'react';
