+++ src/Routes/Routes.jsx
@@ -4,8 +4,8 @@ import Login from '../Pages/Login/Login';
 import Home from '../Pages/Home/Home';
 import Blogs from '../Pages/Blogs/Blogs';
 import BlogDetails from '../Pages/BlogDetails/BlogDetails';
-import Properties from '../Pages/Properties/Properties';
-import PropertiesDetails from '../Components/PropertiesDetails/PropertiesDetails';
+// import Properties from '../Pages/Properties/Properties';
+// import PropertiesDetails from '../Components/PropertiesDetails/PropertiesDetails';
 import Register from '../Pages/Register/Register';
 import Dashboard from '../Layout/Dashboard';
 import HomeDashboard from '../Pages/Dashboard/HomeDashboard/HomeDashboard';
@@ -27,34 +27,36 @@ import PropertyRequest from '../Pages/Dashboard/Moderator/Pages/PropertyRequest/
 import ManagePropertyRequest from '../Pages/Dashboard/Admin/Pages/ManagePropertyRequest/ManagePropertyRequest';
 // import BlogDetails from '../Pages/BlogDetails/BlogDetails';
 import ListingSummary from '../Pages/Dashboard/Admin/Pages/Listing Summary/ListingSummary';
+import Properties from '../Features/Properties/Properties';
+import PropertiesDetails from '../Features/PropertiesDetails/PropertiesDetails';
 // import Payment from '../Components/PaymentMethod/Payment';
 // import ContactUs from '../Components/Contract/ContactUs/ContactUs';
 
 const router = createBrowserRouter([
   {
-    path: "/",
+    path: '/',
     element: <Main></Main>,
     children: [
       {
-        path: "/",
+        path: '/',
         element: <Home />,
       },
       {
-        path: "/blogs",
+        path: '/blogs',
         element: <Blogs />,
       },
       {
-        path: "/blogs/:id",
+        path: '/blogs/:id',
         element: <BlogDetails />,
       },
 
       {
-        path: "/properties",
+        path: '/properties',
         element: <Properties></Properties>,
       },
-      { path: "/createBlog", element: <BlogPost /> },
+      { path: '/createBlog', element: <BlogPost /> },
       {
-        path: "/propertiesDetails/:id",
+        path: '/propertiesDetails/:id',
         element: <PropertiesDetails></PropertiesDetails>,
         loader: ({ params }) =>
           fetch(
@@ -62,53 +64,52 @@ const router = createBrowserRouter([
           ),
       },
       {
-        path: "/addProperties",
+        path: '/addProperties',
         element: <AddProperties />,
       },
       {
-        path: "/login",
+        path: '/login',
         element: <Login />,
       },
       {
-        path: "/register",
+        path: '/register',
         element: <Register />,
       },
       {
         path: '/favorite',
-        element: <FavoriteProperty />
-      }
-
+        element: <FavoriteProperty />,
+      },
     ],
   },
   {
     // dashboard routes
-    path: "/dashboard",
+    path: '/dashboard',
     element: <Dashboard />,
     children: [
       {
-        path: "/dashboard/home",
+        path: '/dashboard/home',
         element: <HomeDashboard />,
       },
       {
-        path: "/dashboard/announcement",
+        path: '/dashboard/announcement',
         element: <Announcement />,
       },
       {
-        path: "/dashboard/all-properties",
+        path: '/dashboard/all-properties',
         element: <AllProperties />,
       },
       {
-        path: "/dashboard/userManagement",
+        path: '/dashboard/userManagement',
         element: <UserManagement />,
       },
       {
-        path: "/dashboard/profile",
+        path: '/dashboard/profile',
         element: <Profile />,
       },
       //? Member only routes
       {
         path: '/dashboard/payment',
-        element: <StripePayment />
+        element: <StripePayment />,
       },
       {
         path: '/dashboard/payment-history',
@@ -116,36 +117,36 @@ const router = createBrowserRouter([
       },
 
       {
-        path: "/dashboard/favourites",
+        path: '/dashboard/favourites',
         element: <Favourites />,
       },
       {
-        path: "/dashboard/property-status",
+        path: '/dashboard/property-status',
         element: <PropertyStatus />,
       },
       {
-        path: "/dashboard/make-announcement",
+        path: '/dashboard/make-announcement',
         element: <MakeAnnouncement />,
       },
       {
-        path: "/dashboard/admin",
+        path: '/dashboard/admin',
         element: <AdminHome />,
       },
       {
-        path: "/dashboard/all-users",
+        path: '/dashboard/all-users',
         element: <AllUsers />,
       },
       {
-        path: "/dashboard/property-request",
+        path: '/dashboard/property-request',
         element: <PropertyRequest />,
       },
       {
-        path: "/dashboard/manage-property-request",
+        path: '/dashboard/manage-property-request',
         element: <ManagePropertyRequest />,
       },
       //? Admin only routes
       {
-        path: "/dashboard/listing-summary",
+        path: '/dashboard/listing-summary',
         element: <ListingSummary />,
       },
     ],
